Moved the database functions to generator, which implements Openmaize.DB behaviour
