+++ lib/mix/openmaize.ex
@@ -2,9 +2,10 @@ defmodule Mix.Openmaize do
   @moduledoc """
   Helper functions for the mix generators.
 
-  There is one mix generator available - `openmaize.gen.phoenixauth`.
-  See the documentation for Mix.Tasks.Openmaize.Gen.Phoenixauth for
-  more information.
+  There are two mix generators available - `openmaize.gen.phoenixauth`
+  and `openmaize.gen.ectodb`.
+  See the documentation for Mix.Tasks.Openmaize.Gen.Phoenixauth and
+  Mix.Tasks.Openmaize.Gen.Ectodb for more information.
   """
 
   @doc """
