+++ priv/templates/database/ecto_db_test.exs
@@ -1,0 +1,7 @@
+defmodule <%= base %>.EctoDBTest do
+  use ExUnit.Case
+  use Plug.Test
+
+  alias <%= base %>.{EctoDB, Repo, User}
+
+end
