+++ test/support/setup_db.exs
@@ -1,6 +1,6 @@
 defmodule Openmaize.SetupDB do
 
-  alias Openmaize.{TestRepo, User}
+  alias Openmaize.{TestRepo, TestUser}
 
   def add_users do
     user1 = %{email: "fred+1@mail.com", username: "fred", role: "user", password: "mangoes&g0oseberries",
@@ -10,23 +10,23 @@ defmodule Openmaize.SetupDB do
     user3 = %{email: "wrong@mail.com", role: "user", password: "mangoes&g0oseberries",
               confirmed_at: nil, confirmation_sent_at: Ecto.DateTime.utc}
 
-    {:ok, _} = %User{}
-    |> User.confirm_changeset(user1, "lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw")
+    {:ok, _} = %TestUser{}
+    |> TestUser.confirm_changeset(user1, "lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw")
     |> TestRepo.insert
-    {:ok, _} = %User{}
-    |> User.confirm_changeset(user2, "lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw")
+    {:ok, _} = %TestUser{}
+    |> TestUser.confirm_changeset(user2, "lg8UXGNMpb5LUGEDm62PrwW8c20qZmIw")
     |> TestRepo.insert
-    {:ok, _} = %User{}
-    |> User.confirm_changeset(user3, "LG9UXGNMpb5LUGEDm62PrwW8c20qZmIw")
+    {:ok, _} = %TestUser{}
+    |> TestUser.confirm_changeset(user3, "LG9UXGNMpb5LUGEDm62PrwW8c20qZmIw")
     |> TestRepo.insert
 
     user4 = %{email: "ray@mail.com", username: "ray", role: "user", password: "h4rd2gU3$$",
               phone: "081555555", confirmed_at: Ecto.DateTime.utc}
-    {:ok, _} = %User{} |> User.auth_changeset(user4) |> TestRepo.insert
+    {:ok, _} = %TestUser{} |> TestUser.auth_changeset(user4) |> TestRepo.insert
 
     user5 = %{email: "brian@mail.com", username: "brian", role: "user", password: "h4rd2gU3$$",
               otp_required: true, otp_secret: "MFRGGZDFMZTWQ2LK"}
-    {:ok, _} = %User{} |> User.auth_changeset(user5) |> TestRepo.insert
+    {:ok, _} = %TestUser{} |> TestUser.auth_changeset(user5) |> TestRepo.insert
   end
 
 end
