+++ test/test_helper.exs
@@ -6,10 +6,10 @@ ExUnit.start()
 
 Code.require_file "support/dummy_crypto.exs", __DIR__
 Code.require_file "support/ecto_helper.exs", __DIR__
+Code.require_file "support/ecto_db.exs", __DIR__
 Code.require_file "support/setup_db.exs", __DIR__
 Code.require_file "support/access_control.exs", __DIR__
 
-Openmaize.SetupDB.add_users()
+Application.put_env(:openmaize, :db_module, Openmaize.EctoDB)
 
-Application.put_env(:openmaize, :repo, Openmaize.TestRepo)
-Application.put_env(:openmaize, :user_model, Openmaize.User)
+Openmaize.SetupDB.add_users()
